+++ DICOM/Network/DicomAssociation.cs
@@ -27,13 +27,11 @@ namespace Dicom.Network
         /// </summary>
         /// <param name="callingAe">The calling Application Entity.</param>
         /// <param name="calledAe">The called Application Entity.</param>
-        /// <param name="maxPduLength">Maximum PDU length.</param>
-        public DicomAssociation(string callingAe, string calledAe, uint maxPduLength = DicomServiceOptions.Default.MaxPDULength)
+        public DicomAssociation(string callingAe, string calledAe)
             : this()
         {
             CallingAE = callingAe;
             CalledAE = calledAe;
-            MaximumPDULength = maxPduLength;
         }
 
         /// <summary>
@@ -77,7 +75,7 @@ namespace Dicom.Network
         public string RemoteImplementationVersion { get; internal set; }
 
         /// <summary>
-        /// Gets the maximum PDU length.
+        /// Gets the maximum PDU length that the remote service accepts.
         /// </summary>
         public uint MaximumPDULength { get; internal set; }
 
@@ -90,7 +88,13 @@ namespace Dicom.Network
         /// Gets the (common) extended negotiations
         /// </summary>
         public DicomExtendedNegotiationCollection ExtendedNegotiations { get; private set; }
-        
+
+
+        /// <summary>
+        /// Gets or sets options to control behavior of <see cref="DicomService"/> base class.
+        /// </summary>
+        public DicomServiceOptions Options { get; internal set; }
+
         /// <summary>
         /// Returns a string that represents the current object.
         /// </summary>
@@ -101,54 +105,70 @@ namespace Dicom.Network
         public override string ToString()
         {
             var sb = new StringBuilder();
-            sb.AppendFormat("Calling AE Title:       {0}\n", CallingAE);
-            sb.AppendFormat("Called AE Title:        {0}\n", CalledAE);
-            sb.AppendFormat("Remote host:            {0}\n", RemoteHost);
-            sb.AppendFormat("Remote port:            {0}\n", RemotePort);
-            sb.AppendFormat(
-                "Implementation Class:   {0}\n",
-                this.RemoteImplementationClassUID ?? DicomImplementation.ClassUID);
-            sb.AppendFormat("Implementation Version: {0}\n", RemoteImplementationVersion ?? DicomImplementation.Version);
-            sb.AppendFormat("Maximum PDU Length:     {0}\n", MaximumPDULength);
-            sb.AppendFormat("Async Ops Invoked:      {0}\n", MaxAsyncOpsInvoked);
-            sb.AppendFormat("Async Ops Performed:    {0}\n", MaxAsyncOpsPerformed);
-            sb.AppendFormat("Presentation Contexts:  {0}\n", PresentationContexts.Count);
+            sb.AppendFormat("Calling AE Title:          {0}\n", CallingAE);
+            sb.AppendFormat("Called AE Title:           {0}\n", CalledAE);
+            sb.AppendFormat("Remote host:               {0}\n", RemoteHost);
+            sb.AppendFormat("Remote port:               {0}\n", RemotePort);
+            sb.AppendFormat("Implementation Class:      {0}\n", RemoteImplementationClassUID ?? DicomImplementation.ClassUID);
+            sb.AppendFormat("Implementation Version:    {0}\n", RemoteImplementationVersion ?? DicomImplementation.Version);
+            sb.AppendFormat("Maximum PDU Length:        {0}\n", MaximumPDULength);
+            sb.AppendFormat("Async Ops Invoked:         {0}\n", MaxAsyncOpsInvoked);
+            sb.AppendFormat("Async Ops Performed:       {0}\n", MaxAsyncOpsPerformed);
+            sb.AppendFormat("Presentation Contexts:     {0}\n", PresentationContexts.Count);
             foreach (var pc in PresentationContexts)
             {
-                sb.AppendFormat("  Presentation Context:  {0} [{1}]\n", pc.ID, pc.Result);
+                sb.AppendFormat("  Presentation Context:     {0} [{1}]\n", pc.ID, pc.Result);
                 if (pc.AbstractSyntax.Name != "Unknown")
-                    sb.AppendFormat("       Abstract Syntax:  {0}\n", pc.AbstractSyntax.Name);
-                else sb.AppendFormat("       Abstract Syntax:  {0}\n", pc.AbstractSyntax);
+                {
+                    sb.AppendFormat("       Abstract Syntax:     {0}\n", pc.AbstractSyntax.Name);
+                }
+                else
+                {
+                    sb.AppendFormat("       Abstract Syntax:     {0}\n", pc.AbstractSyntax);
+                }
+
                 foreach (var tx in pc.GetTransferSyntaxes())
                 {
-                    sb.AppendFormat("       Transfer Syntax:  {0}\n", tx.UID.Name);
+                    sb.AppendFormat("       Transfer Syntax:     {0}\n", tx.UID.Name);
                 }
             }
 
             if (ExtendedNegotiations.Count > 0)
             {
-                sb.AppendFormat("Extended Negotiations:  {0}\n", ExtendedNegotiations.Count);
+                sb.AppendFormat("Extended Negotiations:     {0}\n", ExtendedNegotiations.Count);
                 foreach (DicomExtendedNegotiation ex in ExtendedNegotiations)
                 {
                     if (ex.SopClassUid.Name != "Unknown")
-                        sb.AppendFormat("  Extended Negotiation:  {0}\n", ex.SopClassUid.Name);
-                    else sb.AppendFormat("  Extended Negotiation:  {0}\n", ex.SopClassUid);
+                    {
+                        sb.AppendFormat("  Extended Negotiation:     {0}\n", ex.SopClassUid.Name);
+                    }
+                    else
+                    {
+                        sb.AppendFormat("  Extended Negotiation:     {0}\n", ex.SopClassUid);
+                    }
+
                     if (ex.RequestedApplicationInfo != null)
-                        sb.AppendFormat("      Application Info:  {0}\n", ex.GetApplicationInfo());
+                    {
+                        sb.AppendFormat("      Application Info:     {0}\n", ex.GetApplicationInfo());
+                    }
+
                     if (ex.ServiceClassUid != null)
-                        sb.AppendFormat("         Service Class:  {0}\n", ex.ServiceClassUid);
+                    {
+                        sb.AppendFormat("         Service Class:     {0}\n", ex.ServiceClassUid);
+                    }
+
                     if (ex.RelatedGeneralSopClasses.Any())
                     {
-                        sb.AppendFormat("   Related SOP Classes:  {0}\n", ex.RelatedGeneralSopClasses.Count);
+                        sb.AppendFormat("   Related SOP Classes:     {0}\n", ex.RelatedGeneralSopClasses.Count);
                         foreach (var rel in ex.RelatedGeneralSopClasses)
                         {
-                            sb.AppendFormat("      Related SOP Class:  {0}\n", rel);
+                            sb.AppendFormat("      Related SOP Class:     {0}\n", rel);
                         }
                     }
                 }
             }
 
-            sb.Length = sb.Length - 1;
+            sb.Length -= 1;
             return sb.ToString();
         }
     }