+++ FO-DICOM.Core/Network/DicomService.cs
@@ -457,7 +457,8 @@ namespace FellowOakDicom.Network
                                 Association = new DicomAssociation
                                 {
                                     RemoteHost = _network.RemoteHost,
-                                    RemotePort = _network.RemotePort
+                                    RemotePort = _network.RemotePort,
+                                    Options = Options
                                 };
 
                                 var pdu = new AAssociateRQ(Association);
@@ -689,78 +690,32 @@ namespace FellowOakDicom.Network
                             _dimseStreamFile = null;
 
                             var type = command.GetSingleValue<DicomCommandField>(DicomTag.CommandField);
-                            switch (type)
+                            _dimse = type switch
                             {
-                                case DicomCommandField.CStoreRequest:
-                                    _dimse = new DicomCStoreRequest(command);
-                                    break;
-                                case DicomCommandField.CStoreResponse:
-                                    _dimse = new DicomCStoreResponse(command);
-                                    break;
-                                case DicomCommandField.CFindRequest:
-                                    _dimse = new DicomCFindRequest(command);
-                                    break;
-                                case DicomCommandField.CFindResponse:
-                                    _dimse = new DicomCFindResponse(command);
-                                    break;
-                                case DicomCommandField.CGetRequest:
-                                    _dimse = new DicomCGetRequest(command);
-                                    break;
-                                case DicomCommandField.CGetResponse:
-                                    _dimse = new DicomCGetResponse(command);
-                                    break;
-                                case DicomCommandField.CMoveRequest:
-                                    _dimse = new DicomCMoveRequest(command);
-                                    break;
-                                case DicomCommandField.CMoveResponse:
-                                    _dimse = new DicomCMoveResponse(command);
-                                    break;
-                                case DicomCommandField.CEchoRequest:
-                                    _dimse = new DicomCEchoRequest(command);
-                                    break;
-                                case DicomCommandField.CEchoResponse:
-                                    _dimse = new DicomCEchoResponse(command);
-                                    break;
-                                case DicomCommandField.NActionRequest:
-                                    _dimse = new DicomNActionRequest(command);
-                                    break;
-                                case DicomCommandField.NActionResponse:
-                                    _dimse = new DicomNActionResponse(command);
-                                    break;
-                                case DicomCommandField.NCreateRequest:
-                                    _dimse = new DicomNCreateRequest(command);
-                                    break;
-                                case DicomCommandField.NCreateResponse:
-                                    _dimse = new DicomNCreateResponse(command);
-                                    break;
-                                case DicomCommandField.NDeleteRequest:
-                                    _dimse = new DicomNDeleteRequest(command);
-                                    break;
-                                case DicomCommandField.NDeleteResponse:
-                                    _dimse = new DicomNDeleteResponse(command);
-                                    break;
-                                case DicomCommandField.NEventReportRequest:
-                                    _dimse = new DicomNEventReportRequest(command);
-                                    break;
-                                case DicomCommandField.NEventReportResponse:
-                                    _dimse = new DicomNEventReportResponse(command);
-                                    break;
-                                case DicomCommandField.NGetRequest:
-                                    _dimse = new DicomNGetRequest(command);
-                                    break;
-                                case DicomCommandField.NGetResponse:
-                                    _dimse = new DicomNGetResponse(command);
-                                    break;
-                                case DicomCommandField.NSetRequest:
-                                    _dimse = new DicomNSetRequest(command);
-                                    break;
-                                case DicomCommandField.NSetResponse:
-                                    _dimse = new DicomNSetResponse(command);
-                                    break;
-                                default:
-                                    _dimse = new DicomMessage(command);
-                                    break;
-                            }
+                                DicomCommandField.CStoreRequest => new DicomCStoreRequest(command),
+                                DicomCommandField.CStoreResponse => new DicomCStoreResponse(command),
+                                DicomCommandField.CFindRequest => new DicomCFindRequest(command),
+                                DicomCommandField.CFindResponse => new DicomCFindResponse(command),
+                                DicomCommandField.CGetRequest => new DicomCGetRequest(command),
+                                DicomCommandField.CGetResponse => new DicomCGetResponse(command),
+                                DicomCommandField.CMoveRequest => new DicomCMoveRequest(command),
+                                DicomCommandField.CMoveResponse => new DicomCMoveResponse(command),
+                                DicomCommandField.CEchoRequest => new DicomCEchoRequest(command),
+                                DicomCommandField.CEchoResponse => new DicomCEchoResponse(command),
+                                DicomCommandField.NActionRequest => new DicomNActionRequest(command),
+                                DicomCommandField.NActionResponse => new DicomNActionResponse(command),
+                                DicomCommandField.NCreateRequest => new DicomNCreateRequest(command),
+                                DicomCommandField.NCreateResponse => new DicomNCreateResponse(command),
+                                DicomCommandField.NDeleteRequest => new DicomNDeleteRequest(command),
+                                DicomCommandField.NDeleteResponse => new DicomNDeleteResponse(command),
+                                DicomCommandField.NEventReportRequest => new DicomNEventReportRequest(command),
+                                DicomCommandField.NEventReportResponse => new DicomNEventReportResponse(command),
+                                DicomCommandField.NGetRequest => new DicomNGetRequest(command),
+                                DicomCommandField.NGetResponse => new DicomNGetResponse(command),
+                                DicomCommandField.NSetRequest => new DicomNSetRequest(command),
+                                DicomCommandField.NSetResponse => new DicomNSetResponse(command),
+                                _ => new DicomMessage(command),
+                            };
                             _dimse.PresentationContext =
                                 Association.PresentationContexts.FirstOrDefault(x => x.ID == pdv.PCID);
                             if (!_dimse.HasDataset)