+++ Tests/Desktop/Bugs/VideoCStoreProvider.cs
@@ -2,6 +2,7 @@
 // Licensed under the Microsoft Public License (MS-PL).
 
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
@@ -13,6 +14,8 @@ namespace Dicom.Bugs
 {
     internal class VideoCStoreProvider : DicomService, IDicomServiceProvider, IDicomCStoreProvider
     {
+        private List<string> _storedFiles = new List<string>();
+
         private static readonly DicomTransferSyntax[] AcceptedVideoTransferSyntaxes =
         {
             DicomTransferSyntax.MPEG2,
@@ -46,6 +49,11 @@ namespace Dicom.Bugs
 
         public void OnConnectionClosed(Exception exception)
         {
+            foreach(var tempFile in _storedFiles)
+            {
+                File.Delete(tempFile);
+            }
+            _storedFiles.Clear();
         }
 
         public DicomCStoreResponse OnCStoreRequest(DicomCStoreRequest request)
@@ -53,6 +61,8 @@ namespace Dicom.Bugs
             var tempName = Path.GetTempFileName();
             Logger.Info(tempName);
             request.File.Save(tempName);
+
+            _storedFiles.Add(tempName);
 
             return new DicomCStoreResponse(request, DicomStatus.Success);
         }