+++ Tests/FO-DICOM.Tests/Bugs/GH265.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2012-2020 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
+using FellowOakDicom.Tests.Helpers;
 using System.IO;
 using System.Threading.Tasks;
 using Xunit;
@@ -27,11 +28,14 @@ namespace FellowOakDicom.Tests.Bugs
             var tempName = Path.GetTempFileName();
             DicomFile.Open(TestData.Resolve("CR-MONO1-10-chest")).Save(tempName);
 
-            using var stream = File.OpenRead(tempName);
-            var file = DicomFile.Open(stream);
-            Assert.Equal(DicomFileFormat.DICOM3, file.Format);
-            Assert.True(file.FileMetaInfo.Contains(DicomTag.MediaStorageSOPClassUID));
-            Assert.True(file.FileMetaInfo.Contains(DicomTag.MediaStorageSOPInstanceUID));
+            using (var stream = File.OpenRead(tempName))
+            {
+                var file = DicomFile.Open(stream);
+                Assert.Equal(DicomFileFormat.DICOM3, file.Format);
+                Assert.True(file.FileMetaInfo.Contains(DicomTag.MediaStorageSOPClassUID));
+                Assert.True(file.FileMetaInfo.Contains(DicomTag.MediaStorageSOPInstanceUID));
+            }
+            IOHelper.DeleteIfExists(tempName);
         }
 
         [Fact]
@@ -52,6 +56,8 @@ namespace FellowOakDicom.Tests.Bugs
 
             Assert.Equal(expected.Version, actual.Version);
             Assert.Equal(expected.TransferSyntax.UID.UID, actual.TransferSyntax.UID.UID);
+
+            IOHelper.DeleteIfExists(tempName);
         }
 
         [Fact]
@@ -63,13 +69,16 @@ namespace FellowOakDicom.Tests.Bugs
             var expected = input.FileMetaInfo;
             await input.SaveAsync(tempName).ConfigureAwait(false);
 
-            using var stream = File.OpenRead(tempName);
-            var output = await DicomFile.OpenAsync(stream).ConfigureAwait(false);
-            var actual = output.FileMetaInfo;
+            using (var stream = File.OpenRead(tempName))
+            {
+                var output = await DicomFile.OpenAsync(stream).ConfigureAwait(false);
+                var actual = output.FileMetaInfo;
 
-            Assert.NotEqual(expected.ImplementationClassUID.UID, actual.ImplementationClassUID.UID);
-            Assert.NotEqual(expected.ImplementationVersionName, actual.ImplementationVersionName);
-            Assert.Equal(expected.SourceApplicationEntityTitle, actual.SourceApplicationEntityTitle);
+                Assert.NotEqual(expected.ImplementationClassUID.UID, actual.ImplementationClassUID.UID);
+                Assert.NotEqual(expected.ImplementationVersionName, actual.ImplementationVersionName);
+                Assert.Equal(expected.SourceApplicationEntityTitle, actual.SourceApplicationEntityTitle);
+            }
+            IOHelper.DeleteIfExists(tempName);
         }
 
         [Fact]