+++ ChangeLog.md
@@ -6,6 +6,7 @@
 * Optimize DicomTag.GetHashCode() (#1088)
 * Bug fix: Prevent special characters in association requests from crashing Fellow Oak DICOM (#1104)
 * Fixes caching of pixel data in DicomImage when rendering the same image multithreaded. (#805)
+* Fix handling of MaximumPDULength. DicomServer always repeated the clients value of MaximumPDULength in AssociationAccepted-message instead of returning its own value. (#1084)
 
 #### v.4.0.6 (8/6/2020)
 * Update to DICOM Standard 2020b.